refactor(S-3.02): extract handle_schemas+types+schema to assets/schemas.rs
Moves handle_schemas, handle_types, handle_schema, resolve_schema,
ambiguous_type_error, format_attribute_type + all resolve_schema and
format_attribute_type unit tests to schemas.rs shard. mod.rs is now a
pure dispatch layer. All 612 unit tests pass.
